requirements.txt
streamlit
pandas
numpy
gspread
oauth2client
google-api-python-client
python-dotenv